server/requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.8
gevent==20.5.0
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
torch==1.1.0
torchvision==0.3.0
urllib3==1.25.8
Werkzeug==1.0.1
